requirements.txt
altgraph==0.17.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
CTkToolTip==0.8
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.11
numpy==2.4.2
packaging==26.0
pep8==1.7.1
pillow==12.1.0
psutil==7.2.2
pulsectl==24.12.0
pycparser==3.0
pygame==2.6.1
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2026.0
pyttsx3==2.99
requests==2.32.5
setuptools==80.10.2
sounddevice==0.5.5
soundfile==0.13.1
tkinter-tooltip==3.1.2
urllib3==2.6.3
